requirements.txt
click
kafka-python
